+++ dbms/src/Storages/Page/ExternalPageCallbacks.h
@@ -1,0 +1,38 @@
+// Copyright 2022 PingCAP, Ltd.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#pragma once
+
+#include <Storages/Page/PageDefines.h>
+
+#include <vector>
+
+namespace DB
+{
+
+struct ExternalPageCallbacks
+{
+    // `scanner` for scanning available external page ids on disks.
+    // `remover` will be called with living normal page ids after gc run a round, user should remove those
+    //           external pages(files) in `pending_external_pages` but not in `valid_normal_pages`
+    using PathAndIdsVec = std::vector<std::pair<String, std::set<PageId>>>;
+    using ExternalPagesScanner = std::function<PathAndIdsVec()>;
+    using ExternalPagesRemover
+        = std::function<void(const PathAndIdsVec & pending_external_pages, const std::set<PageId> & valid_normal_pages)>;
+    ExternalPagesScanner scanner = nullptr;
+    ExternalPagesRemover remover = nullptr;
+    NamespaceId ns_id = MAX_NAMESPACE_ID;
+};
+
+} // namespace DB
