+++ dbms/src/Storages/Page/PageStorage.h
@@ -19,6 +19,7 @@
 #include <Interpreters/SettingsCommon.h>
 #include <Storages/FormatVersion.h>
 #include <Storages/Page/Config.h>
+#include <Storages/Page/ExternalPageCallbacks.h>
 #include <Storages/Page/FileUsage.h>
 #include <Storages/Page/Page.h>
 #include <Storages/Page/PageDefines.h>
@@ -64,20 +65,6 @@ enum class PageStorageRunMode : UInt8
     ONLY_V2 = 1,
     ONLY_V3 = 2,
     MIX_MODE = 3,
-};
-
-struct ExternalPageCallbacks
-{
-    // `scanner` for scanning available external page ids on disks.
-    // `remover` will be called with living normal page ids after gc run a round, user should remove those
-    //           external pages(files) in `pending_external_pages` but not in `valid_normal_pages`
-    using PathAndIdsVec = std::vector<std::pair<String, std::set<PageId>>>;
-    using ExternalPagesScanner = std::function<PathAndIdsVec()>;
-    using ExternalPagesRemover
-        = std::function<void(const PathAndIdsVec & pending_external_pages, const std::set<PageId> & valid_normal_pages)>;
-    ExternalPagesScanner scanner = nullptr;
-    ExternalPagesRemover remover = nullptr;
-    NamespaceId ns_id = MAX_NAMESPACE_ID;
 };
 
 /**
